UART: 32-bit numbers fix
UART send num was uint8_t and numbers bigger than 255 was not printed right.

uart_send_num() was changed to get uint32_t num parameter.

+++ uart/uart.c
@@ -43,7 +43,7 @@ void uart_send_string(char *text)
     }
 }
 
-void uart_send_num(uint8_t num, uint8_t radix)
+void uart_send_num(uint32_t num, uint8_t radix)
 {
 	itoa(num, num_str, radix);
 	uart_send_string(num_str);
